Use existing abstraction for hard-coded string.
We were interpolating a hard-coded string, which makes no sense.
Either the string is hard-coded and is part of the template, or it's
not. In this case, even though it doesn't vary explicitly, we do have
an abstraction for it. It's the aside for the very first critter that
the little old lady swallows.